cdc-platform/requirements.txt
# Web Framework
Flask==3.0.0
gunicorn==21.2.0
Flask-CORS==4.0.0

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Redis
redis==5.0.1

# Kafka
kafka-python==2.0.2
confluent-kafka==2.3.0

# Data Validation
pydantic==2.5.2
pydantic-settings==2.1.0

# Security
PyJWT==2.8.0
cryptography==41.0.7
bcrypt==4.1.2

# HTTP Client
requests==2.31.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
ulid-py==1.1.0

# Monitoring
prometheus-client==0.19.0
